requirements/local.txt
# django-debug-toolbar that works with Django 1.5+
django-debug-toolbar==1.3.2
